requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
tzdata==2024.2

